--- a/jest-fast.config.js
+++ b/jest-fast.config.js
@@ -4,7 +4,11 @@ module.exports = {
         "^.+\\.(t|j)sx?$": "@swc/jest",
     },
     testEnvironment: "node",
-    testPathIgnorePatterns: ["/node_modules/", "/dist/"],
+    testPathIgnorePatterns: [
+        "/node_modules/",
+        "/dist/",
+        "/src/test/e2e-emulated/map*",
+    ],
     maxWorkers: "8",
     globalSetup: "./jest.setup.js",
     globalTeardown: "./jest.teardown.js",