@anton-trunov 2025-01-16
Here is the test performance difference observed on my machine for `yarn test`:

### Before this PR

```shell
Test Suites: 341 passed, 341 total
Tests:       25 skipped, 6255 passed, 6280 total
Snapshots:   742 passed, 742 total
Time:        30.752 s, estimated 56 s
Ran all test suites.
✨  Done in 31.57s.
yarn test  359.74s user 30.41s system 1226% cpu 31.822 total
```

### After this PR

```shell
Test Suites: 341 passed, 341 total
Tests:       25 skipped, 6255 passed, 6280 total
Snapshots:   742 passed, 742 total
Time:        18.423 s, estimated 30 s
Ran all test suites.
✨  Done in 19.36s.
yarn test  169.37s user 13.46s system 934% cpu 19.572 total
```